#563 | refactor(masonry/types): move Size, Point, and Bounds to common.types
