+++ modules/masonry/src/components/Brick/BrickWidget.tsx
@@ -1,8 +1,9 @@
 import { useState } from 'react';
+
 import type { WidgetInput } from '@/@types/brick.types';
 
-import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from '@/ui/select';
 import { Input } from '@/ui/input';
+import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from '@/ui/select';
 import { Slider } from '@/ui/slider';
 
 export interface WidgetProps {
