+++ modules/masonry/src/components/Path/Path.stories.tsx
@@ -1,4 +1,5 @@
 import type { Meta, StoryObj } from '@storybook/react-vite';
+
 import type { BrickOutlineInput } from '@/@types/brick.types';
 
 import { PathBrickView } from './Path';
